refactor:(backend) ♻️ reorganize modules structure
