+++ backend/src/module/app/app.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ConfigModule } from '@nestjs/config';
-import { UserModule } from 'src/user/user.module';
-import { AuthModule } from 'src/auth/auth.module';
+import { UserModule } from 'src/module/user/user.module';
+import { AuthModule } from 'src/module/auth/auth.module';
 
 @Module({
   imports: [
