+++ backend/src/module/auth/dto/sign-in-response.dto.ts
