+++ backend/src/module/auth/entities/auth.entity.ts
