+++ backend/src/module/auth/repository/prisma-auth.repository.ts
@@ -1,5 +1,5 @@
 import { Injectable } from "@nestjs/common";
-import { PrismaService } from "src/prisma/prisma.service";
+import { PrismaService } from "src/module/prisma/prisma.service";
 import { hash, compare } from 'bcrypt';
 import { RefreshTokenMapper } from "../mappers/auth.mapper";
