+++ backend/src/module/user/entities/enums/seniority.enum.ts
