+++ backend/src/module/user/user.module.ts
@@ -3,7 +3,7 @@ import { UserService } from './user.service';
 import { UserController } from './user.controller';
 import { UserRepository } from './repository/user.repository';
 import { PrismaRepository } from './repository/prisma-user.repository';
-import { PrismaModule } from 'src/prisma/prisma.module';
+import { PrismaModule } from 'src/module/prisma/prisma.module';
 
 @Module({
   imports: [PrismaModule],
